+++ src/Layers/xrRender/r__screenshot.cpp
@@ -76,29 +76,33 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 					NULL, pSrcSmallTexture ));
 #endif
 
-				// save (logical & physical)
-				ID3DBlob*		saved	= 0;
-#ifdef USE_DX11
-				HRESULT hr = D3DX11SaveTextureToMemory(HW.pContext, pSrcSmallTexture, D3DX11_IFF_DDS, &saved, 0);
+			// save (logical & physical)
+#if defined(USE_DX10) || defined(USE_DX11)
+			ID3D10Blob*			saved	= 0; // D3DX10/11SaveTextureToMemory (7d migrates to DirectXTex)
 #else
-				HRESULT hr					= D3DX10SaveTextureToMemory( pSrcSmallTexture, D3DX10_IFF_DDS, &saved, 0);
-				//HRESULT hr					= D3DXSaveTextureToFileInMemory (&saved,D3DXIFF_DDS,texture,0);
+			ID3DXBuffer*		saved	= 0; // D3DXSave*ToFileInMemory (7d migrates to DirectXTex)
 #endif
-				if(hr==D3D_OK)
+#ifdef USE_DX11
+			HRESULT hr = D3DX11SaveTextureToMemory(HW.pContext, pSrcSmallTexture, D3DX11_IFF_DDS, &saved, 0);
+#else
+			HRESULT hr					= D3DX10SaveTextureToMemory( pSrcSmallTexture, D3DX10_IFF_DDS, &saved, 0);
+			//HRESULT hr					= D3DXSaveTextureToFileInMemory (&saved,D3DXIFF_DDS,texture,0);
+#endif
+			if(hr==D3D_OK)
+			{
+				IWriter*			fs		= FS.w_open	(name); 
+				if (fs)				
 				{
-					IWriter*			fs		= FS.w_open	(name); 
-					if (fs)				
-					{
-						fs->w				(saved->GetBufferPointer(),(u32)saved->GetBufferSize());
-						FS.w_close			(fs);
-					}
+					fs->w				(saved->GetBufferPointer(),(u32)saved->GetBufferSize());
+					FS.w_close			(fs);
 				}
-				_RELEASE			(saved);
-
-				// cleanup
-				_RELEASE			(pSrcSmallTexture);
 			}
-			break;
+			_RELEASE			(saved);
+
+			// cleanup
+			_RELEASE			(pSrcSmallTexture);
+		}
+		break;
 		case IRender_interface::SM_FOR_MPSENDING:
 			{
 				
@@ -125,13 +129,17 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 				CHK_DX(D3DX10LoadTextureFromTexture( pSrcTexture,
 					NULL, pSrcSmallTexture ));
 #endif
-				// save (logical & physical)
-				ID3DBlob*		saved	= 0;
-#ifdef USE_DX11
-				HRESULT hr	= D3DX11SaveTextureToMemory(HW.pContext, pSrcSmallTexture, D3DX11_IFF_DDS, &saved, 0);
+			// save (logical & physical)
+#if defined(USE_DX10) || defined(USE_DX11)
+			ID3D10Blob*			saved	= 0;
 #else
-				HRESULT hr					= D3DX10SaveTextureToMemory( pSrcSmallTexture, D3DX10_IFF_DDS, &saved, 0);
-				//HRESULT hr					= D3DXSaveTextureToFileInMemory (&saved,D3DXIFF_DDS,texture,0);
+			ID3DXBuffer*		saved	= 0;
+#endif
+#ifdef USE_DX11
+			HRESULT hr	= D3DX11SaveTextureToMemory(HW.pContext, pSrcSmallTexture, D3DX11_IFF_DDS, &saved, 0);
+#else
+			HRESULT hr					= D3DX10SaveTextureToMemory( pSrcSmallTexture, D3DX10_IFF_DDS, &saved, 0);
+			//HRESULT hr					= D3DXSaveTextureToFileInMemory (&saved,D3DXIFF_DDS,texture,0);
 #endif
 				if(hr==D3D_OK)
 				{
@@ -159,8 +167,12 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 			{
 				string64			t_stemp;
 				string_path			buf;
-				xr_sprintf			(buf,sizeof(buf),"ss_%s_%s_(%s).jpg",Core.UserName,timestamp(t_stemp),(g_pGameLevel)?g_pGameLevel->name().c_str():"mainmenu");
-				ID3DBlob			*saved	= 0;
+			xr_sprintf			(buf,sizeof(buf),"ss_%s_%s_(%s).jpg",Core.UserName,timestamp(t_stemp),(g_pGameLevel)?g_pGameLevel->name().c_str():"mainmenu");
+#if defined(USE_DX10) || defined(USE_DX11)
+			ID3D10Blob			*saved	= 0;
+#else
+			ID3DXBuffer			*saved	= 0;
+#endif
 #ifdef USE_DX11
 				CHK_DX				(D3DX11SaveTextureToMemory(HW.pContext, pSrcTexture, D3DX11_IFF_JPG, &saved, 0));
 #else
@@ -173,8 +185,12 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 
 				if (strstr(Core.Params,"-ss_tga"))	
 				{ // hq
-					xr_sprintf			(buf,sizeof(buf),"ssq_%s_%s_(%s).tga",Core.UserName,timestamp(t_stemp),(g_pGameLevel)?g_pGameLevel->name().c_str():"mainmenu");
-					ID3DBlob*		saved	= 0;
+				xr_sprintf			(buf,sizeof(buf),"ssq_%s_%s_(%s).tga",Core.UserName,timestamp(t_stemp),(g_pGameLevel)?g_pGameLevel->name().c_str():"mainmenu");
+#if defined(USE_DX10) || defined(USE_DX11)
+				ID3D10Blob*			saved	= 0;
+#else
+				ID3DXBuffer*		saved	= 0;
+#endif
 #ifdef USE_DX11
 					CHK_DX				(D3DX11SaveTextureToMemory(HW.pContext, pSrcTexture, D3DX11_IFF_BMP, &saved, 0));
 #else
@@ -312,7 +328,7 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 				if(hr!=D3D_OK)		goto _end_;
 
 				// save (logical & physical)
-				ID3DBlob*		saved	= 0;
+				ID3DXBuffer*		saved	= 0;
 				hr					= D3DXSaveTextureToFileInMemory (&saved,D3DXIFF_DDS,texture,0);
 				if(hr!=D3D_OK)		goto _end_;
 				
@@ -345,7 +361,7 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 				if(hr!=D3D_OK)		goto _end_;
 
 				// save (logical & physical)
-				ID3DBlob*		saved	= 0;
+				ID3DXBuffer*		saved	= 0;
 				hr					= D3DXSaveTextureToFileInMemory (&saved,D3DXIFF_DDS,texture,0);
 				if(hr!=D3D_OK)		goto _end_;
 				
@@ -372,7 +388,7 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 				string64			t_stemp;
 				string_path			buf;
 				xr_sprintf			(buf,sizeof(buf),"ss_%s_%s_(%s).jpg",Core.UserName,timestamp(t_stemp),(g_pGameLevel)?g_pGameLevel->name().c_str():"mainmenu");
-				ID3DBlob*		saved	= 0;
+				ID3DXBuffer*		saved	= 0;
 				CHK_DX				(D3DXSaveSurfaceToFileInMemory (&saved,D3DXIFF_JPG,pFB,0,0));
 				IWriter*		fs	= FS.w_open	("$screenshots$",buf); R_ASSERT(fs);
 				fs->w				(saved->GetBufferPointer(),saved->GetBufferSize());
@@ -380,7 +396,7 @@ void CRender::ScreenshotImpl	(ScreenshotMode mode, LPCSTR name, CMemoryWriter* m
 				_RELEASE			(saved);
 				if (strstr(Core.Params,"-ss_tga"))	{ // hq
 					xr_sprintf			(buf,sizeof(buf),"ssq_%s_%s_(%s).tga",Core.UserName,timestamp(t_stemp),(g_pGameLevel)?g_pGameLevel->name().c_str():"mainmenu");
-					ID3DBlob*		saved	= 0;
+					ID3DXBuffer*		saved	= 0;
 					CHK_DX				(D3DXSaveSurfaceToFileInMemory (&saved,D3DXIFF_TGA,pFB,0,0));
 					IWriter*		fs	= FS.w_open	("$screenshots$",buf); R_ASSERT(fs);
 					fs->w				(saved->GetBufferPointer(),saved->GetBufferSize());